verifier speedup via wNAF powering
Normally, wNAF doesn't make sense for RSA because inversion is
expensive. But in this case, we have to compute the inverse of
C1 and C2 anyway, meaning that we can use Montgomery's trick to
get inverses of Aq, Bq, Cq essentially for free. That lets us
use signed exponent recoding to make pow2 much faster.